test: split schematic graphic cases
